--- providers/microsoft.keyvault/controls/KeyVault_Errors.json
@@ -1,5 +1,0 @@
-{
-"ControlId": "KeyVault_AdvancedCheck",
-"Category": "EXAMPLE: Authentication strength, Attack surface reduction",
-"Description": "EXAMPLE: Ensure The Service calls downstream resources with managed identity"
-}
